+++ web-server/plugins/slycat-parameter-image/js/selectors.ts
